feat: errorHandler.js 추가
